+++ src/Sample/SMS.Fingerprint.Sample/MainView.xaml.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Threading;
+using System.Threading.Tasks;
+using Plugin.Fingerprint.Abstractions;
 using Xamarin.Forms;
 
 namespace SMS.Fingerprint.Sample
@@ -19,6 +21,27 @@ namespace SMS.Fingerprint.Sample
             lblStatus.Text = "";
             var result = await Plugin.Fingerprint.CrossFingerprint.Current.AuthenticateAsync("Prove you have fingers!", _cancel.Token);
 
+            await SetResultAsync(result);
+        }
+
+        private async void OnAuthenticateLocalized(object sender, EventArgs e)
+        {
+            _cancel = swAutoCancel.IsToggled ? new CancellationTokenSource(TimeSpan.FromSeconds(10)) : new CancellationTokenSource();
+            lblStatus.Text = "";
+
+            var dialogConfig = new DialogConfiguration("Beweise, dass du Finger hast!")
+            {
+                CancelTitle = "Abbrechen",
+                FallbackTitle = "Anders!"
+            };
+
+            var result = await Plugin.Fingerprint.CrossFingerprint.Current.AuthenticateAsync(dialogConfig, _cancel.Token);
+
+            await SetResultAsync(result);
+        }
+
+        private async Task SetResultAsync(FingerprintAuthenticationResult result)
+        {
             if (result.Authenticated)
             {
                 await Navigation.PushAsync(new SecretView());