+++ doc/changelog.md
@@ -2,5 +2,10 @@
 
 ### 1.3.0
 
+#### 1.3.0-beta2
+- #13 Fallback button on Android
+- #12 possibility to set localized fallback title
+- #6 fixed crash on Android
+
 #### 1.3.0-beta1
 - #5 Implemented feedback for failed authentication try